+++ app/src/main/java/com/example/mynewsapplication/data/response/NewsDataFromJson.kt
@@ -1,7 +1,5 @@
 package com.example.mynewsapplication.data.response
 
-import com.example.mynewsapplication.data.response.Article
-
 data class NewsDataFromJson(
     val articles: List<Article>,
     val status: String,
