+++ app/src/main/java/com/example/mynewsapplication/data/response/Source.kt
@@ -1,9 +1,6 @@
 package com.example.mynewsapplication.data.response
 
-import com.example.mynewsapplication.data.response.Article
-
 data class Source(
     val id: Any,
-    val name: String,
-    val sameArticles: MutableList<Article>
+    val name: String
 )
