+++ app/src/main/java/com/example/mynewsapplication/presentation/SourceFragment.kt
@@ -6,9 +6,11 @@ import android.widget.Toast
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.mynewsapplication.base.BaseFragment
+import com.example.mynewsapplication.data.repository.NewsRepository
 import com.example.mynewsapplication.databinding.FragmentSourceBinding
 import com.example.mynewsapplication.model.NewsModel
 import com.example.mynewsapplication.presentation.adapter.SourceAdapter
+import com.example.mynewsapplication.viewmodel.SourceViewModel
 
 class SourceFragment(private val newsDataList: MutableList<NewsModel>, private val newsFilterer: (String) -> Unit) : BaseFragment<FragmentSourceBinding>(){
     override fun inflateBinding(
@@ -18,19 +20,20 @@ class SourceFragment(private val newsDataList: MutableList<NewsModel>, private v
         return FragmentSourceBinding.inflate(inflater, container, false)
     }
 
+    private lateinit var sourceViewModel: SourceViewModel
     override fun setupView() {
+        sourceViewModel = SourceViewModel(NewsRepository())
+
         val recyclerView : RecyclerView = binding.recyclerView
         recyclerView.layoutManager =
             LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
-        var uniqueDataSource: MutableList<String> = getUniqueSource(newsDataList)
+        var uniqueDataSource: MutableList<String> = sourceViewModel.getUniqueSource(newsDataList)
         val adapter = SourceAdapter(uniqueDataSource)
         recyclerView.adapter = adapter
 
         adapter.setOnItemClickListener(object : SourceAdapter.OnItemClickListener{
             override fun onItemClick(position: Int) {
                 Toast.makeText(context, uniqueDataSource[position], Toast.LENGTH_SHORT).show()
-//                val fragment = ArticleFragment(newsDataList, uniqueDataSource[position])
-
                 newsFilterer.invoke(uniqueDataSource[position])
                 parentFragmentManager.popBackStack()
             }
@@ -46,15 +49,5 @@ class SourceFragment(private val newsDataList: MutableList<NewsModel>, private v
         adapter.setOnItemLongClickListener(object : SourceAdapter.OnItemLongClickListener {
             override fun onItemLongClick(position: Int) = Unit
         })
-    }
-
-    fun getUniqueSource(newsList: MutableList<NewsModel>): MutableList<String> {
-        val uniqueSources = HashSet<String>()
-
-        for (news in newsList) {
-            news.source?.let { uniqueSources.add(it) }
-        }
-
-        return uniqueSources.toMutableList()
     }
 }
