+++ app/src/main/java/com/example/mynewsapplication/presentation/adapter/FragmentAdapter.kt
@@ -5,8 +5,6 @@ import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
 
-
-
 class FragmentAdapter(fm: FragmentManager, lifecycle: Lifecycle) : FragmentStateAdapter(fm, lifecycle){
 
     private val fragments: MutableList<Fragment> = mutableListOf()
