+++ app/src/main/java/com/example/mynewsapplication/viewmodel/NewsViewModel.kt
@@ -1,6 +1,5 @@
 package com.example.mynewsapplication.viewmodel
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.mynewsapplication.model.NewsModel
