+++ app/src/main/java/com/example/mynewsapplication/viewmodel/SourceViewModel.kt
@@ -1,0 +1,11 @@
+package com.example.mynewsapplication.viewmodel
+
+import androidx.lifecycle.ViewModel
+import com.example.mynewsapplication.data.repository.NewsRepository
+import com.example.mynewsapplication.model.NewsModel
+
+class SourceViewModel(private val repository: NewsRepository): ViewModel() {
+    fun getUniqueSource(newsList: MutableList<NewsModel>): MutableList<String>{
+        return repository.getUniqueSource(newsList)
+    }
+}
